0.0.19 full transaction handling - use locale from env variable LANG - use correct username - revise DbResponse, esp. with respect to multi-valued responses - revised Connection a bit, due to revised DbResponse (now we have: query(), dml(), execute(), what_ever()) - fixed some FIXMEs - added ymd factory to LongDate

+++ src/protocol/lowlevel/parts/client_info.rs
@@ -23,8 +23,8 @@ impl ClientInfo {
         }
         len
     }
-    pub fn count(&self) -> i16 {
-        self.0.len() as i16
+    pub fn count(&self) -> usize {
+        self.0.len()
     }
 
     pub fn parse(no_of_args: i32, rdr: &mut BufRead) -> PrtResult<ClientInfo> {
